backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
sentence-transformers==5.2.0
faiss-cpu==1.12.0
numpy==2.3.5
google-generativeai==0.8.6
pydantic==2.9.0
